# Conflicts:
#	devtools/devtools-integration/ohos/src/devtools_napi.cc
#	docs/development/native-integration.md
#	docs/development/native-module.md
#	dom/include/dom/dom_manager.h
#	dom/include/dom/dom_node.h
#	dom/include/dom/layout_node.h
#	dom/src/dom/animation/animation_manager.cc
#	dom/src/dom/dom_manager.cc
#	dom/src/dom/root_node.cc
#	driver/js/examples/hippy-vue-demo/src/components/native-demos/demo-waterfall.vue
#	driver/js/include/driver/napi/js_ctx.h
#	driver/js/include/driver/napi/jsh/jsh_ctx.h
#	driver/js/include/driver/scope.h
#	driver/js/src/napi/jsh/jsh_ctx.cc
#	driver/js/src/scope.cc
#	framework/examples/ohos-demo/hvigorfile.ts
#	framework/examples/ohos-demo/obfuscation-rules.txt
#	framework/examples/ohos-demo/src/main/cpp/CMakeLists.txt
#	framework/examples/ohos-demo/src/main/ets/hippy_extend/ExampleNativeTurboModule.ets
#	framework/examples/ohos-demo/src/main/ets/hippy_extend/ExampleViewA.ets
#	framework/examples/ohos-demo/src/main/ets/hippy_extend/turbo/TurboConfig.ets
#	framework/examples/ohos-demo/src/main/ets/pages/PageManagement.ets
#	framework/examples/ohos-har-demo/hvigorfile.ts
#	framework/examples/ohos-har-demo/obfuscation-rules.txt
#	framework/examples/ohos-har-demo/src/main/ets/pages/Index.ets
#	framework/ohos/build-profile.json5
#	framework/ohos/hvigorfile.ts
#	framework/ohos/index.ets
#	framework/ohos/obfuscation-rules.txt
#	framework/ohos/src/main/cpp/CMakeLists.txt
#	framework/ohos/src/main/cpp/impl/CMakeLists.txt
#	framework/ohos/src/main/cpp/impl/connector/CMakeLists.txt
#	framework/ohos/src/main/cpp/impl/connector/include/connector/arkts_turbo_module.h
#	framework/ohos/src/main/cpp/impl/connector/include/connector/js2ark.h
#	framework/ohos/src/main/cpp/impl/connector/include/connector/turbo.utils.h
#	framework/ohos/src/main/cpp/impl/connector/src/arkts_turbo_module.cc
#	framework/ohos/src/main/cpp/impl/connector/src/dom_manager_napi.cc
#	framework/ohos/src/main/cpp/impl/connector/src/js2ark.cc
#	framework/ohos/src/main/cpp/impl/connector/src/js_driver_napi.cc
#	framework/ohos/src/main/cpp/impl/connector/src/setting_napi.cc
#	framework/ohos/src/main/cpp/impl/connector/src/turbo_module_manager.cc
#	framework/ohos/src/main/cpp/impl/connector/src/turbo_utils.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/CMakeLists.txt
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/arkui/arkui_node.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/arkui/text_node.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/components/base_view.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/components/rich_text_image_span_view.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/components/rich_text_span_view.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/components/rich_text_view.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/native_render_context.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/native_render_manager.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/native_render_provider.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/uimanager/hr_view_manager.h
#	framework/ohos/src/main/cpp/impl/renderer/native/include/renderer/utils/hr_pixel_utils.h
#	framework/ohos/src/main/cpp/impl/renderer/native/src/arkui/arkui_node.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/arkui/text_node.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/components/base_view.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/components/rich_text_image_span_view.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/components/rich_text_span_view.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/components/rich_text_view.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/native_render_manager.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/native_render_provider.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/native_render_provider_capi.cc
#	framework/ohos/src/main/cpp/impl/renderer/native/src/native_render_provider_napi.cc
#	framework/ohos/src/main/ets/hippy_framework/HippyEngine.ets
#	framework/ohos/src/main/ets/hippy_framework/HippyEngineContextImpl.ets
#	framework/ohos/src/main/ets/hippy_framework/HippyEngineManagerImpl.ets
#	framework/ohos/src/main/ets/hippy_framework/connector/DomManager.ets
#	framework/ohos/src/main/ets/hippy_framework/connector/JsDriver.ets
#	framework/ohos/src/main/ets/hippy_framework/devsupport/DevRemoteServerData.ets
#	framework/ohos/src/main/ets/hippy_framework/index.ets
#	framework/ohos/src/main/ets/hippy_library/HippyLibrary.ets
#	framework/ohos/src/main/ets/renderer_native/index.ets
#	framework/ohos/src/main/ets/support/annotation/HippyNativeModule.ets
#	framework/ohos/src/main/ets/support/utils/PixelUtil.ets
#	hvigorfile.ts
#	modules/footstone/include/footstone/cv_driver.h
#	modules/footstone/include/footstone/driver.h
#	modules/footstone/include/footstone/string_view_utils.h
#	modules/footstone/include/footstone/task_runner.h
#	modules/footstone/src/serializer.cc
#	modules/footstone/src/task_runner.cc
#	modules/footstone/src/worker.cc
#	modules/ohos/oh_napi/CMakeLists.txt
#	renderer/native/ios/renderer/component/view/HippyView.m

+++ framework/ios/base/executors/HippyJSEnginesMapper.mm
@@ -60,7 +60,15 @@ EngineResource::EngineResource(const std::string name) {
 }
 
 EngineResource::~EngineResource() {
-    dom_worker_->Terminate();
+    auto runner = engine_->GetJsTaskRunner();
+    if (footstone::Worker::IsTaskRunning() && runner == footstone::runner::TaskRunner::GetCurrentTaskRunner()) {
+        __block auto dom_worder = dom_worker_;
+        dispatch_async(dispatch_get_main_queue(), ^{
+            dom_worder->Terminate();
+        });
+    } else {
+        dom_worker_->Terminate();
+    }
 }
 
 using EngineRef = std::pair<std::shared_ptr<EngineResource>, NSUInteger>;
